apps/backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
openai==1.12.0
boto3==1.34.69
flask-sqlalchemy==3.1.1
pymysql==1.1.0